Corrected file pathing for componets that were updated and moved. File structure needed correct syntax
