+++ src/Components/AuthForm/LogInForm.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import * as userService from '../../utilities/users-service'
+import * as userService from '../../utilities/users-utilities/users-service'
 
 export default function LogInForm ({setUser}) {
