+++ src/utilities/users-utilities/users-api.js
@@ -1,4 +1,4 @@
-import { sendRequest } from "../send-request";
+import sendRequest from "../send-request";
 const BASE_URL = "/api/users";
 
 // user data from sign up form, passed through
